# app/openapi.yaml
openapi: 3.1.0
info:
  title: AI Risk Control Agent for MT5
  version: "1.0.0"
  description: |
    HTTP surface for the MT5 risk-analysis service.

    Two analysis paths:
      - `POST /analyse_risk` — caller supplies a snapshot and gets per-risk verdicts
        synchronously (Phase A passthrough).
      - `POST /run_scan` — server pulls a 6h window from Alex, buckets by login,
        builds historical context, analyses, persists, and posts the batch to the
        configured callback (Phase B).

    Stored verdicts can be retrieved with `GET /analyses`.
servers:
  - url: http://localhost:5000
    description: Local development

tags:
  - name: health
  - name: analysis
  - name: scan

paths:
  /healthz:
    get:
      tags: [health]
      summary: Liveness probe
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }

  /analyse_risk:
    post:
      tags: [analysis]
      summary: Analyse a caller-supplied snapshot
      description: |
        Runs all configured risks against the provided snapshot. Re-POSTs for the
        same `(mt5_login, start_time)` overwrite the previously stored rows.
        The callback URL is invoked with the same body fire-and-forget.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountSnapshot"
      responses:
        "200":
          description: Analysis complete
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RiskAnalysis" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }

  /run_scan:
    post:
      tags: [scan]
      summary: Trigger a manual scan for a 6h window
      description: |
        Pulls all activity for the requested window from Alex, runs the full
        analysis pipeline for every login present, persists the verdicts, and
        delivers a single batched callback. If no body is supplied the server
        defaults to the most recently completed 6h window aligned to
        `00/06/12/18 UTC`.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                start_time:
                  type: string
                  format: date-time
                  example: "2026-05-09T00:00:00Z"
                end_time:
                  type: string
                  format: date-time
                  example: "2026-05-09T05:59:59.999Z"
              required: [start_time, end_time]
      responses:
        "200":
          description: Scan completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanResult" }
        "400":
          description: Window arguments invalid or partial
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "502":
          description: Upstream Alex fetch failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ScanResult" }

  /analyses:
    get:
      tags: [analysis]
      summary: Fetch persisted verdicts for a login + window start
      parameters:
        - in: query
          name: mt5_login
          required: true
          schema: { type: integer, example: 1000123 }
        - in: query
          name: start_time
          required: true
          schema:
            type: string
            format: date-time
            example: "2026-05-09T00:00:00Z"
      responses:
        "200":
          description: Stored rows for the window
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StoredAnalysis" }
        "400":
          description: Missing or malformed parameters
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: No analyses found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /openapi.yaml:
    get:
      tags: [health]
      summary: This OpenAPI document (raw YAML)
      responses:
        "200":
          description: OpenAPI spec
          content:
            application/yaml: {}

  /docs:
    get:
      tags: [health]
      summary: Swagger UI for this API
      responses:
        "200":
          description: HTML page
          content:
            text/html: {}

components:
  responses:
    ValidationError:
      description: Body failed schema validation
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: validation_error }
              details:
                type: array
                items: { type: object }
    UnsupportedMediaType:
      description: Body must be application/json
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }

  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      additionalProperties: true

    Trade:
      type: object
      required:
        [
          id,
          login,
          group,
          symbol,
          volume,
          side,
          open_time,
          time,
          open_price,
          close_price,
          bid_at_open,
          ask_at_open,
          profit,
        ]
      properties:
        id: { type: integer }
        login: { type: integer }
        group: { type: string }
        entry: { type: integer, default: 0, description: "0=in, 1=out" }
        symbol: { type: string }
        volume: { type: number, format: float }
        side: { type: string, enum: [buy, sell] }
        open_time: { type: string, format: date-time }
        time:
          type: string
          format: date-time
          description: "Close time (alias serialised as `close_time`)."
        open_price: { type: number }
        close_price: { type: number }
        bid_at_open: { type: number }
        ask_at_open: { type: number }
        stop_loss: { type: number, default: 0 }
        take_profit: { type: number, default: 0 }
        swaps: { type: number, default: 0 }
        commission: { type: number, default: 0 }
        profit: { type: number }

    Deposit:
      type: object
      required: [id, login, group, time, profit]
      properties:
        id: { type: integer }
        login: { type: integer }
        group: { type: string }
        time: { type: string, format: date-time }
        profit: { type: number }

    Withdraw:
      allOf:
        - $ref: "#/components/schemas/Deposit"

    Bonus:
      allOf:
        - $ref: "#/components/schemas/Deposit"

    AccountSnapshot:
      type: object
      required:
        [mt5_login, start_time, end_time, trades, deposits, withdraws, bonus]
      properties:
        mt5_login: { type: integer, example: 1000123 }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        trades:
          type: array
          items: { $ref: "#/components/schemas/Trade" }
        deposits:
          type: array
          items: { $ref: "#/components/schemas/Deposit" }
        withdraws:
          type: array
          items: { $ref: "#/components/schemas/Withdraw" }
        bonus:
          type: array
          items: { $ref: "#/components/schemas/Bonus" }

    SubRuleEvaluation:
      type: object
      required: [rule, observed_value, "true", reason]
      properties:
        rule: { type: string }
        observed_value: {}
        "true": { type: boolean }
        reason: { type: string }

    RiskAnalysis:
      type: object
      required:
        [
          mt5_login,
          risk_type,
          risk_score,
          risk_level,
          evidence,
          suggested_action,
          analysis,
          start_time,
          end_time,
          trigger_type,
        ]
      properties:
        mt5_login: { type: integer }
        risk_type:
          type: string
          enum: [latency_arbitrage, scalping, swap_arbitrage, bonus_abuse]
        risk_score: { type: integer, minimum: 0, maximum: 100 }
        risk_level:
          type: string
          enum: [none, watch, medium, high, critical]
        evidence:
          type: object
          properties:
            evaluations:
              type: array
              items: { $ref: "#/components/schemas/SubRuleEvaluation" }
            summary: { type: string }
        suggested_action: { type: string }
        analysis: { type: string }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        trigger_type:
          type: string
          enum: [scheduled_scan, manual_run, api_call]

    StoredAnalysis:
      allOf:
        - $ref: "#/components/schemas/RiskAnalysis"

    ScanResult:
      type: object
      required:
        [
          start_time,
          end_time,
          accounts_analysed,
          accounts_failed,
          results_persisted,
          error,
          failed_logins,
        ]
      properties:
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        accounts_analysed: { type: integer }
        accounts_failed: { type: integer }
        results_persisted: { type: integer }
        error:
          type: [string, "null"]
          description: null on success; short code on failure (e.g. "alex_fetch_failed:...").
        failed_logins:
          type: array
          items: { type: integer }